+++ lib/libedit/filecomplete.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: filecomplete.h,v 1.8 2009/02/16 00:15:45 christos Exp $	*/
+/*	$NetBSD: filecomplete.h,v 1.6 2008/04/29 06:53:01 martin Exp $	*/
 
 /*-
  * Copyright (c) 1997 The NetBSD Foundation, Inc.
@@ -34,10 +34,10 @@
 int fn_complete(EditLine *,
     char *(*)(const char *, int),
     char **(*)(const char *, int, int),
-    const char *, const char *, const char *(*)(const char *), size_t,
+    const char *, const char *, const char *(*)(const char *), int,
     int *, int *, int *, int *);
 
-void fn_display_match_list(EditLine *, char **, size_t, size_t);
+void fn_display_match_list(EditLine *, char **, int, int);
 char *fn_tilde_expand(const char *);
 char *fn_filename_completion_function(const char *, int);
